fix(common/es_clients): move it into common/es/clients.py
